Move readers, converters, utils to their own packages Made SQLite classes inner classes of SQLiteReader
