+++ app/src/main/java/com/quantimodo/etl/converters/CardiographConverter.java
@@ -1,7 +1,9 @@
-package com.quantimodo.etl;
+package com.quantimodo.etl.converters;
 
 import com.quantimodo.android.sdk.model.Measurement;
 import com.quantimodo.android.sdk.model.MeasurementSet;
+import com.quantimodo.etl.DatabaseView;
+import com.quantimodo.etl.Table;
 
 import java.util.ArrayList;
