+++ app/src/main/java/com/quantimodo/etl/readers/Reader.java
@@ -1,4 +1,6 @@
-package com.quantimodo.etl;
+package com.quantimodo.etl.readers;
+
+import com.quantimodo.etl.DatabaseView;
 
 import java.io.File;
 import java.io.IOException;
