+++ app/src/main/java/com/quantimodo/etl/readers/SpreadsheetReader.java
@@ -1,4 +1,8 @@
-package com.quantimodo.etl;
+package com.quantimodo.etl.readers;
+
+import com.quantimodo.etl.ArrayDatabaseView;
+import com.quantimodo.etl.ArrayTable;
+import com.quantimodo.etl.DatabaseView;
 
 import org.apache.poi.openxml4j.exceptions.InvalidFormatException;
 import org.apache.poi.ss.formula.eval.NotImplementedException;
